winelayer/pyproject.toml
[project]
name = "winelayer-daemon"
version = "0.1.0"
description = "WineLayer — Intelligent Windows app compatibility layer for Linux"
requires-python = ">=3.11"
license = {text = "LGPL-2.1-or-later"}

dependencies = [
    "sqlalchemy>=2.0",
    "aiosqlite>=0.19",
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
